feat: support partial authentication (#372)
Given a path with a GET endpoint that should be public and a PUT endpoint that should be protected,
it is now possible to register two separate notarized routes, one within the 'authenticate' block and one outside.
Also had to split up TestModules because precommit complained about it being too long.